docs: add `.know-token` endpoint for local development
Implement a new `.know-token` endpoint to facilitate API testing in the
local environment. This endpoint allows developers to quickly obtain a
valid ID Token for authentication and testing of the various APIs.

The new endpoint is designed with security in mind, ensuring that users
can only access tokens generated for their own authenticated account, and
not those of other users.

Detailed instructions on how to use this new endpoint are documented in
the authentication guide: [Authentication](./docs/rest/authentication.md).